Move custom exception handler to exceptions.py
