+++ api/base/utils.py
@@ -55,18 +55,3 @@ def waterbutler_url_for(request_type, provider, path, node_id, token, obj_args=N
 
     url.args.update(query)
     return url.url
-
-
-def custom_exception_handler(exc, context):
-    """
-    Custom exception handler that nests detail inside errors.
-    """
-    from rest_framework.views import exception_handler
-    response = exception_handler(exc, context)
-
-    if response is not None:
-        if 'detail' in response.data:
-            response.data = {'errors': [response.data]}
-        else:
-            response.data = {'errors': [{'detail': response.data}]}
-    return response
